refactor: rename path-access to path-helpers and drop unused exports
